+++ apps/staking/pages/index.tsx
@@ -14,6 +14,7 @@ import {
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 import { TermOfServices, Container, MavaWidget } from "ui-helpers";
+import { greyCliff, ibm, nb } from "ui-helpers/src/fonts";
 import { MixpanelProvider } from "tracker";
 function SnackbarsInternal() {
   const valueRedux = useSelector((state: StoreType) => getAllSnackbars(state));
@@ -25,6 +26,7 @@ import { HeadComponent } from "../src/components/staking/HeadComponent";
 import { GoogleAnalytics } from "../src/components/GoogleAnalytics";
 import { StatefulFooter } from "stateful-components";
 import { GiveFeedback } from "../src/GiveFeedback";
+import { cn } from "helpers";
 
 const Content = dynamic(() => import("../src/components/staking/Content"));
 
@@ -41,7 +43,9 @@ export default function Home() {
             <>
               <HeadComponent />
               <GoogleAnalytics />
-              <main>
+              <main
+                className={cn(greyCliff.variable, ibm.variable, nb.variable)}
+              >
                 <TermOfServices />
                 <GiveFeedback />
                 <Container>