+++ packages/ui-helpers/src/TopBarItem.tsx
@@ -17,9 +17,11 @@ export const TopBarItem = ({
   const drawDiv = () => (
     <div>
       <div className="flex items-center justify-center space-x-2">
-        <h5 className="text-sm font-normal opacity-80">{text}</h5>
+        <h5 className="text-sm font-normal opacity-80 ">{text}</h5>
       </div>
-      <h2 className="font-[GreyCliff] text-2xl font-bold">{value}</h2>
+      <h2 className="font-body text-2xl font-bold flex justify-center">
+        {value}
+      </h2>
     </div>
   );