+++ packages/ui-helpers/src/index.tsx
@@ -73,3 +73,4 @@ export { Spinner } from "./spinners/Spinner";
 export * from "./Toggle";
 export * from "./selects/SelectMenu";
 export * from "./modal/WizardHelper";
+export { PrimaryLink } from "./links/Link";