+++ packages/ui-helpers/src/termsOfServices/TermOfServices.tsx
@@ -44,14 +44,14 @@ export const TermOfServices = () => {
     setConsent(!consent);
     setShowConsent(true);
     setModalContent(
-      <ConsentModal setShow={setShowConsent} setConsent={setConsent} />,
+      <ConsentModal setShow={setShowConsent} setConsent={setConsent} />
     );
   };
 
   return (
     <ModalTOS title="Evmos Terms of Service" show={show}>
       <div className="space-y-3">
-        <div className="border-darkGray5 h-80 w-full space-y-3 overflow-y-auto border p-4 font-[IBM]">
+        <div className="border-darkGray5 h-80 w-full space-y-3 overflow-y-auto border p-4 font-display">
           <Content />
         </div>
         <div className="space-y-3">